restructure script task for easier test, add a unittest
