+++ packages/nexrender-core/test/fixtures/paramInject.jsx
@@ -4,8 +4,7 @@ TODO: should not pick up params in comments
 // TODO: should not pick up params in comments
 // TODO: should/does it pick up params in quoted strings?
 
-const undefined_param = NX.get('undefined_parameter')
-const defined_param = NX.get('defined_parameter')
-
+const a = NX.get('unittest_undefined_parameter')
+const b = NX.get('unittest_defined_parameter')
 
 const result = NX.call('onePlusOne', 1)
